pyproject.toml
[tool.poetry]
name = "muni"
version = "0.1.0"
description = ""
authors = ["orhan <orhancansu98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
django-filter = "^21.1"
django-taggit = "^1.5.1"
django-rest-auth = "^0.9.5"
djangorestframework-jwt = "^1.11.0"
drf-spectacular = "^0.20.2"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
isort = "^5.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
